AdminEvent: Enable Delete Event
Removed the old method of deleting an event by manually inserting an
event ID and enabled it through actions in the events table. Also added
a small delete check to make sure the user wants to delete the event. I
also moved and edited the code for the delete notification to display
the name of the event that was deleted.

DeleteEventModal.png
https://patina-dev.nyc3.digitaloceanspaces.com/screenshot/Screenshot-2024-11-13T23:40:30-MartinM-DeleteEventModal.png

DeleteEventNotif.png
 https://patina-dev.nyc3.digitaloceanspaces.com/screenshot/Screenshot-2024-11-13T23:41:18-MartinM-DeleteEventNotif.png

TEST=manual
Ran page locally to verify feature works.
